+++ routes/adminroutes.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, request, jsonify
 import os
 import requests
+from decorators import role_required
 from database.admindatahandler import is_admin
 from database.userdatahandler import get_images_by_user, get_recent_uploads, get_upload_stats
 from utils.clerk_auth import require_auth
@@ -10,7 +11,7 @@ admin_bp = Blueprint('admin', __name__, url_prefix='/api/admin')
 
 # Get all images uploaded by a user (admin access)
 @admin_bp.route('/user_uploads/<user_id>')
-@require_auth
+@role_required('admin')
 def admin_user_images_show(user_id):
     try:
         images = get_images_by_user(user_id)
@@ -24,7 +25,7 @@ def admin_user_images_show(user_id):
 
 # Get all users
 @admin_bp.route('/users', methods=['GET'])
-@require_auth
+@role_required('admin')
 def get_users():
     try:
         # Get query parameters
@@ -74,7 +75,7 @@ def get_users():
 
 # Get only users (not admins)
 @admin_bp.route('/users/only-users', methods=['GET'])
-@require_auth
+@role_required('admin')
 def get_only_users():
     try:
         # Get query parameters
@@ -127,7 +128,7 @@ def get_only_users():
 
 # Get dashboard statistics and recent activity
 @admin_bp.route('/dashboard', methods=['GET'])
-@require_auth
+@role_required('admin')
 def get_dashboard_data():
     try:
         # Get query parameters for recent activity
